Force data tracking to reconnect to CDM properties when loading a new state.
Circuit and Graph components can vary between states, and we don't have capacity to try to preserve common objects between all configurations, so we just delete those objects and create new ones based on the configuration in the loaded state. This means the engine tracker is holding dead pointers to those deleted objects, so we need to rehook them up to the new circuit/compartment objects.